requirements.txt
torch>=2.5
torchvision>=0.20
tqdm>=4.66
pandas>=2.2
matplotlib>=3.9
scikit-learn>=1.5
pillow>=10.0
